camera refactor
